+++ src/main/scala/groundtest/BusMasterTest.scala
@@ -4,7 +4,7 @@ import Chisel._
 import uncore.tilelink._
 import uncore.agents._
 import uncore.coherence.{InnerTLId, OuterTLId}
-import uncore.util._
+import util._
 import junctions.HasAddrMapParameters
 import cde.Parameters
