+++ src/main/scala/groundtest/Regression.scala
@@ -4,9 +4,8 @@ import Chisel._
 import uncore.tilelink._
 import uncore.constants._
 import uncore.agents._
-import uncore.util._
+import util._
 import junctions.HasAddrMapParameters
-import util.{ParameterizedBundle, SimpleTimer}
 import rocket.HellaCacheIO
 import cde.{Parameters, Field}
