+++ src/main/scala/rocket/btb.scala
@@ -4,10 +4,8 @@ package rocket
 
 import Chisel._
 import cde.{Parameters, Field}
-import Util._
-import uncore.util._
+import util._
 import uncore.agents.PseudoLRU
-import util.ParameterizedBundle
 
 case object BtbKey extends Field[BtbParameters]
