+++ src/main/scala/rocket/frontend.scala
@@ -2,7 +2,7 @@ package rocket
 
 import Chisel._
 import uncore.tilelink._
-import Util._
+import util._
 import cde.{Parameters, Field}
 
 class FrontendReq(implicit p: Parameters) extends CoreBundle()(p) {
