+++ src/main/scala/rocket/ibuf.scala
@@ -3,9 +3,8 @@
 package rocket
 
 import Chisel._
-import Util._
+import util._
 import cde.{Parameters, Field}
-import util.ParameterizedBundle
 
 class Instruction(implicit val p: Parameters) extends ParameterizedBundle with HasCoreParameters {
   val pf0 = Bool() // page fault on first half of instruction
