+++ src/main/scala/rocket/icache.scala
@@ -4,7 +4,7 @@ import Chisel._
 import uncore.agents._
 import uncore.tilelink._
 import uncore.util._
-import Util._
+import util._
 import cde.{Parameters, Field}
 
 trait HasL1CacheParameters extends HasCacheParameters with HasCoreParameters {
