+++ src/main/scala/rocket/multiplier.scala
@@ -4,7 +4,7 @@ package rocket
 
 import Chisel._
 import ALU._
-import Util._
+import util._
 
 class MultiplierReq(dataBits: Int, tagBits: Int) extends Bundle {
   val fn = Bits(width = SZ_ALU_FN)
