+++ src/main/scala/rocket/rocc.scala
@@ -6,7 +6,7 @@ import Chisel._
 import uncore.tilelink._
 import uncore.constants._
 import uncore.agents.CacheName
-import Util._
+import util._
 import cde.{Parameters, Field}
 
 case object RoccMaxTaggedMemXacts extends Field[Int]
