+++ src/main/scala/rocket/rvc.scala
@@ -2,9 +2,8 @@ package rocket
 
 import Chisel._
 import Chisel.ImplicitConversions._
-import Util._
+import util._
 import cde.Parameters
-import uncore.util._
 
 class ExpandedInstruction extends Bundle {
   val bits = UInt(width = 32)
