+++ src/main/scala/rocketchip/Periphery.scala
@@ -12,7 +12,7 @@ import uncore.converters._
 import uncore.devices._
 import uncore.agents._
 import uncore.util._
-import rocket.Util._
+import util._
 import rocket.XLen
 import scala.math.max
 import coreplex._
