+++ src/main/scala/uncore/agents/Cache.scala
@@ -11,6 +11,7 @@ import uncore.coherence._
 import uncore.tilelink._
 import uncore.constants._
 import uncore.util._
+import util._
 import cde.{Parameters, Field}
 
 case object CacheName extends Field[String]
