+++ src/main/scala/util/BlackBoxRegs.scala
@@ -1,4 +1,4 @@
-package uncore.util
+package util
 
 import Chisel._
